requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper==20231117
anthropic==0.28.0
pydantic==2.7.1